feat: add following tab logic
