--- app/pages/home/following.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-definePageMeta({
-  layout: 'home',
-});
-</script>
-
-<template>
-  <div></div>
-</template>
